chore: Move tests into tests directory
