+++ crates/pithos_lib/src/model/serialization.rs
@@ -321,45 +321,3 @@ impl RecipientSection {
         Ok(())
     }
 }
-
-// --- Tests ---
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_varint_encoding() {
-        use integer_encoding::VarIntWriter;
-        let mut buf = Vec::new();
-        buf.write_varint(0u64).unwrap();
-        assert_eq!(buf, vec![0x00]);
-        buf.clear();
-        buf.write_varint(127u64).unwrap();
-        assert_eq!(buf, vec![0x7F]);
-        buf.clear();
-        buf.write_varint(128u64).unwrap();
-        assert_eq!(buf, vec![0x80, 0x01]);
-        buf.clear();
-        buf.write_varint(300u64).unwrap();
-        assert_eq!(buf, vec![0xAC, 0x02]);
-    }
-
-    #[test]
-    fn test_string_encoding() {
-        let s = "hello";
-        let mut buf = Vec::new();
-        encode_string(&mut buf, s).unwrap();
-        assert_eq!(buf, vec![5, b'h', b'e', b'l', b'l', b'o']);
-    }
-
-    #[test]
-    fn test_file_header_serialization() {
-        let mut buf = Vec::new();
-        let fh = FileHeader {
-            magic: *b"PITH",
-            version: 0x0100,
-        };
-        fh.serialize(&mut buf).unwrap();
-        assert_eq!(buf, vec![b'P', b'I', b'T', b'H', 0x01, 0x00]);
-    }
-}
